Add support for prop overrides
The format of the overrides is KEY=VALUE
Where KEY is one of the dict keys used in config
https://cs.android.com/android/platform/superproject/main/+/main:build/soong/scripts/gen_build_prop.py
And VALUE is the value you want take over.
VALUE can't contain spaces, in such cases you can use normalize-path-list to normalize it,
replacing spaces with a double colon.

An example:
PRODUCT_BUILD_PROP_OVERRIDES := \
    DeviceName=jfltevzw \
    DeviceProduct=jfltevzw \
    SystemDevice=jfltevzw \
    SystemName=jfltevzw \
    BuildFingerprint=Verizon/jfltevzw/jfltevzw:4.2.2/JDQ39/I545VRUAMDK:user/release-keys \
    BuildDesc=$(call normalize-path-list, "jfltevzw-user 4.2.2 JDQ39 I545VRUAMDK release-keys")

Change-Id: Ib44a3bb573d08d492acc092e104cb4f687f3b168